baseboard: Simple baseboard for PyCom development modules
This board does not use SMD circuits. Instead low price modules for the USB/UART
bridge and the SD card adapter are used. The USB/UART bridge is the red FTDI
module, which are sold at prices of $1.5 on. The SD card module is the one of
Pololu, which is also widely available. Pictures of the modules are included
in the repository. Depending on the soldered connectors, the board can also
be used for WiPy1. Artwork on the board indicates, how the Pycom modules have
to be inserted.
The board also carries the two transistor logic used by espressif to switch into
bootstrap mode. The Pycom command line updater supports that with the -r option.